requirements.txt
PyYAML==6.0.2
aiomqtt==2.3.0
homeassistant>=2025.2.0
